feat: Context에 훅 병합, 변수명 직관적으로 수정
