+++ src/components/BookListItem.jsx
@@ -1,11 +1,11 @@
-import { useCurrentBook } from '@/hooks/useCurrentBook'
-import { useSearch } from '@/hooks/useSearchBook'
+import { useBook } from '@/context/BookContext'
+import { useSearch } from '@/context/SearhContext'
 import { memo } from 'react'
 import { Link } from 'react-router-dom'
 
 const BookListItem = () => {
   console.log('[BookItem] -rerender')
-  const { savedCurrentBook } = useCurrentBook()
+  const { saveBook } = useBook()
   const { filteredBooks } = useSearch()
 
   return (
@@ -19,7 +19,7 @@ const BookListItem = () => {
           </Link>
           <button
             style={{ padding: '0.2rem 0.4rem', marginLeft: 4 }}
-            onClick={() => savedCurrentBook({ title: book.title })}
+            onClick={() => saveBook({ title: book.title })}
           >
             읽기
           </button>
