+++ src/components/Header.jsx
@@ -1,5 +1,4 @@
 const Header = () => {
-  // console.log('[Header] -rerender') 리렌더링 안 됨
   return <div style={{ backgroundColor: '#888' }}>책을 골라봅시다</div>
 }
